refactor(ipv4): Use byte-arrays to store addresses and u16s
On the network, all data is encoded as big-endian. When using data-types
such as u16 or u32, their internal encoding depends on the endianness
of the host machine.

In order to make reasoning about the data within those fields easier, we
change them to byte-arrays and encode/decode from/to network endianness
directly.

Related: #32